+++ crates/client/src/commander.rs
@@ -19,6 +19,8 @@ use tokio::{
 };
 
 use crate::constants::*;
+use crate::fuzzing_stats::FuzzingStatistics;
+use tokio::io::AsyncBufReadExt;
 
 #[derive(Error, Debug)]
 pub enum Error {
@@ -153,7 +155,11 @@ impl Commander {
         // arguments so we need to parse the variable content.
         let hfuzz_run_args = std::env::var("HFUZZ_RUN_ARGS").unwrap_or_default();
 
-        let fuzz_args = config.get_honggfuzz_args(hfuzz_run_args);
+        let rustflags = std::env::var("RUSTFLAGS").unwrap_or_default();
+
+        let rustflags = config.get_rustflags_args(rustflags);
+
+        let mut fuzz_args = config.get_honggfuzz_args(hfuzz_run_args);
 
         // let cargo_target_dir = std::env::var("CARGO_TARGET_DIR").unwrap_or_default();
 
@@ -181,36 +187,34 @@ impl Commander {
             }
         }
 
-        let mut rustflags = if config.fuzz.allow_duplicate_txs {
-            "--cfg allow_duplicate_txs "
-        } else {
-            ""
-        }
-        .to_string();
-
-        rustflags.push_str(&std::env::var("RUSTFLAGS").unwrap_or_default());
-
-        let mut child = Command::new("cargo")
-            .env("HFUZZ_RUN_ARGS", fuzz_args)
-            .env("CARGO_TARGET_DIR", cargo_target_dir)
-            .env("HFUZZ_WORKSPACE", hfuzz_workspace)
-            .env("RUSTFLAGS", rustflags)
-            .arg("hfuzz")
-            .arg("run")
-            .arg(target)
-            .spawn()?;
-
-        tokio::select! {
-            res = child.wait() =>
-                match res {
-                    Ok(status) => if !status.success() {
-                        println!("Honggfuzz exited with an error!");
-                    },
-                    Err(_) => throw!(Error::FuzzingFailed),
-            },
-            _ = signal::ctrl_c() => {
-                tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
-            },
+        match rustflags.contains("fuzzing_with_stats") {
+            true => {
+                // enforce keep output to be true
+                fuzz_args.push_str("--keep_output");
+                let mut child = Command::new("cargo")
+                    .env("HFUZZ_RUN_ARGS", fuzz_args)
+                    .env("CARGO_TARGET_DIR", cargo_target_dir)
+                    .env("HFUZZ_WORKSPACE", hfuzz_workspace)
+                    .env("RUSTFLAGS", rustflags)
+                    .arg("hfuzz")
+                    .arg("run")
+                    .arg(target)
+                    .stdout(Stdio::piped())
+                    .spawn()?;
+                Self::handle_child_with_stats(&mut child).await?;
+            }
+            false => {
+                let mut child = Command::new("cargo")
+                    .env("HFUZZ_RUN_ARGS", fuzz_args)
+                    .env("CARGO_TARGET_DIR", cargo_target_dir)
+                    .env("HFUZZ_WORKSPACE", hfuzz_workspace)
+                    .env("RUSTFLAGS", rustflags)
+                    .arg("hfuzz")
+                    .arg("run")
+                    .arg(target)
+                    .spawn()?;
+                Self::handle_child(&mut child).await?;
+            }
         }
 
         if let Ok(crash_files) = get_crash_files(&crash_dir, &ext) {
@@ -236,27 +240,54 @@ impl Commander {
         let hfuzz_workspace = std::env::var("HFUZZ_WORKSPACE")
             .unwrap_or_else(|_| config.get_env_arg("HFUZZ_WORKSPACE"));
 
-        let fuzz_args = config.get_honggfuzz_args(hfuzz_run_args);
+        let mut fuzz_args = config.get_honggfuzz_args(hfuzz_run_args);
 
-        let mut rustflags = if config.fuzz.allow_duplicate_txs {
-            "--cfg allow_duplicate_txs "
-        } else {
-            ""
+        let rustflags = std::env::var("RUSTFLAGS").unwrap_or_default();
+
+        let rustflags = config.get_rustflags_args(rustflags);
+
+        match rustflags.contains("fuzzing_with_stats") {
+            true => {
+                // enforce keep output to be true
+                fuzz_args.push_str("--keep_output");
+                let mut child = Command::new("cargo")
+                    .env("HFUZZ_RUN_ARGS", fuzz_args)
+                    .env("CARGO_TARGET_DIR", cargo_target_dir)
+                    .env("HFUZZ_WORKSPACE", hfuzz_workspace)
+                    .env("RUSTFLAGS", rustflags)
+                    .arg("hfuzz")
+                    .arg("run")
+                    .arg(target)
+                    .stdout(Stdio::piped())
+                    .spawn()?;
+                Self::handle_child_with_stats(&mut child).await?;
+            }
+            false => {
+                let mut child = Command::new("cargo")
+                    .env("HFUZZ_RUN_ARGS", fuzz_args)
+                    .env("CARGO_TARGET_DIR", cargo_target_dir)
+                    .env("HFUZZ_WORKSPACE", hfuzz_workspace)
+                    .env("RUSTFLAGS", rustflags)
+                    .arg("hfuzz")
+                    .arg("run")
+                    .arg(target)
+                    .spawn()?;
+                Self::handle_child(&mut child).await?;
+            }
         }
-        .to_string();
+    }
 
-        rustflags.push_str(&std::env::var("RUSTFLAGS").unwrap_or_default());
-
-        let mut child = Command::new("cargo")
-            .env("HFUZZ_RUN_ARGS", fuzz_args)
-            .env("CARGO_TARGET_DIR", cargo_target_dir)
-            .env("HFUZZ_WORKSPACE", hfuzz_workspace)
-            .env("RUSTFLAGS", rustflags)
-            .arg("hfuzz")
-            .arg("run")
-            .arg(target)
-            .spawn()?;
-
+    /// Manages a child process in an async context, specifically for monitoring fuzzing tasks.
+    /// Waits for the process to exit or a Ctrl+C signal. Prints an error message if the process
+    /// exits with an error, and sleeps briefly on Ctrl+C. Throws `Error::FuzzingFailed` on errors.
+    ///
+    /// # Arguments
+    /// * `child` - A mutable reference to a `Child` process.
+    ///
+    /// # Errors
+    /// * Throws `Error::FuzzingFailed` if waiting on the child process fails.
+    #[throws]
+    async fn handle_child(child: &mut Child) {
         tokio::select! {
             res = child.wait() =>
                 match res {
@@ -267,6 +298,69 @@ impl Commander {
             },
             _ = signal::ctrl_c() => {
                 tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
+            },
+        }
+    }
+    /// Asynchronously manages a child fuzzing process, collecting and logging its statistics.
+    /// This function spawns a new task dedicated to reading the process's standard output and logging the fuzzing statistics.
+    /// It waits for either the child process to exit or a Ctrl+C signal to be received. Upon process exit or Ctrl+C signal,
+    /// it stops the logging task and displays the collected statistics in a table format.
+    ///
+    /// The implementation ensures that the statistics logging task only stops after receiving a signal indicating the end of the fuzzing process
+    /// or an interrupt from the user, preventing premature termination of the logging task if scenarios where reading is faster than fuzzing,
+    /// which should not be common.
+    ///
+    /// # Arguments
+    /// * `child` - A mutable reference to a `Child` process, representing the child fuzzing process.
+    ///
+    /// # Errors
+    /// * `Error::FuzzingFailed` - Thrown if there's an issue with managing the child process, such as failing to wait on the child process.
+    #[throws]
+    async fn handle_child_with_stats(child: &mut Child) {
+        let stdout = child
+            .stdout
+            .take()
+            .expect("child did not have a handle to stdout");
+
+        let reader = tokio::io::BufReader::new(stdout);
+
+        let fuzz_end = std::sync::Arc::new(std::sync::atomic::AtomicBool::new(false));
+        let fuzz_end_clone = std::sync::Arc::clone(&fuzz_end);
+
+        tokio::spawn(async move {
+            let mut stats_logger = FuzzingStatistics::new();
+
+            let mut lines = reader.lines();
+            loop {
+                // Why the lock ??
+                // I`m not sure what happens if the fuzzing sessions is still active
+                // however the reader already read all the lines. I think that would bring us
+                // to the scenario where this reading task ends prematurely.
+                if fuzz_end_clone.load(std::sync::atomic::Ordering::SeqCst) {
+                    break;
+                }
+                while let Ok(Some(line)) = lines.next_line().await {
+                    stats_logger.insert_serialized(&line);
+                }
+            }
+            stats_logger.show_table();
+        });
+
+        tokio::select! {
+            res = child.wait() =>{
+                fuzz_end.store(true, std::sync::atomic::Ordering::SeqCst);
+                match res {
+                    Ok(status) => {
+                        if !status.success() {
+                            println!("Honggfuzz exited with an error!");
+                        }
+                    },
+                    Err(_) => throw!(Error::FuzzingFailed),
+                }
+            },
+            _ = signal::ctrl_c() => {
+                tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
+                fuzz_end.store(true, std::sync::atomic::Ordering::SeqCst);
             },
         }
     }
@@ -286,14 +380,9 @@ impl Commander {
         let cargo_target_dir = std::env::var("CARGO_TARGET_DIR")
             .unwrap_or_else(|_| config.get_env_arg("CARGO_TARGET_DIR"));
 
-        let mut rustflags = if config.fuzz.allow_duplicate_txs {
-            "--cfg allow_duplicate_txs "
-        } else {
-            ""
-        }
-        .to_string();
+        let rustflags = std::env::var("RUSTFLAGS").unwrap_or_default();
 
-        rustflags.push_str(&std::env::var("RUSTFLAGS").unwrap_or_default());
+        let rustflags = config.get_rustflags_args(rustflags);
 
         // using exec rather than spawn and replacing current process to avoid unflushed terminal output after ctrl+c signal
         std::process::Command::new("cargo")